+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 1.3.1 (Status: DEVELOPMENT)
+Version 1.3.1 (Status: RELEASED)
 * Bugfix: Slaves reported incorrect date to nafsdmctl slavestatus (minute instead of month under the "latest connection date" section)
 * Slaves now check if another nafsdm-slave process is running before boot
 * Fix issue #22